ANN Update #3: More Tests, Negative Weights
Added the possibility for negative weights and biases in the random initialization of the neural network. Added an initialization test program. Combined the Turtle Sigmoid Test and Derivative Test.
